Given a string, find the longest length of palindrome starting from the left. For example: 	abaeab, the longest from left is aba, so return 3. 	abcbaef, the longest from left is abcba, so return 5. 	ababa, the longest from left is ababa, so return 5.
The idea is from calculating next in KMP.
Take abae for example, let’s add the reverse to the right: abaeeaba.
Let’s implement KMP, set i=0, j=4. Then, let’s find the max next_value.
We can find the max next_value in O(n) time. In this example, the max next_value is 3. And that is the result.