*Given n boxes with length and width. A box with smaller length and smaller width can be put on a box with larger length and larger width.  Each box height is 1. Find the tallest box that they can put together.
For n boxes, we do comparison between each box. If box[j] can be put on box[i], then we put an edge from box[i]->box[j].
In this way, we build a graph among boxes. Gathering all the boxes which doesn't has parent, we use BFS to iterate.
The max number of iteration will be the result.