https://leetcode.com/problems/flatten-nested-list-iterator/ This problem is the same as deepIterator. Each time, we store the iterator of the list in stack. But in my previous deepIterator, I didn't handle when a list is empty.
For this problem, I use currInteger to tell if iterator has next.

The main part is the refresh() function. It iterates until it finds an element is Integer. When current iterator doesn't has next or list is totally empty, it pops the stop. When current element is a list, it pushes to stack.